requirements.txt
-i https://pypi.org/simple
annotated-types==0.5.0; python_version >= '3.7'
anyio==3.7.1; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
exceptiongroup==1.1.3; python_version < '3.11'
fastapi==0.103.1; python_version >= '3.7'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
h11==0.14.0; python_version >= '3.7'
httpcore==0.18.0; python_version >= '3.8'
httptools==0.6.0
httpx==0.25.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
libcst==1.0.1
markdown-it-py==3.0.0; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==23.1; python_version >= '3.7'
pluggy==1.3.0; python_version >= '3.8'
pydantic==2.3.0; python_version >= '3.7'
pydantic-core==2.6.3; python_version >= '3.7'
pygments==2.16.1
pytest==7.4.2; python_version >= '3.7'
pytest-asyncio==0.21.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1; python_version >= '3.6'
rich==13.5.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
strawberry-graphql[debug-server]==0.208.1; python_version >= '3.8' and python_version < '4.0'
tomli==2.0.1; python_version < '3.11'
typer==0.9.0
typing-extensions==4.7.1; python_version >= '3.7'
typing-inspect==0.9.0
uvicorn[standard]==0.23.2; python_version >= '3.8'
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
